# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tennis Ranya API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type LoginInputRole = typeof LoginInputRole[keyof typeof LoginInputRole];


export const LoginInputRole = {
  admin: 'admin',
  cashier: 'cashier',
} as const;

export interface LoginInput {
  username: string;
  password: string;
  role: LoginInputRole;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  cashier: 'cashier',
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
}

export interface AuthResult {
  user: User;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  cashier: 'cashier',
} as const;

export interface UserInput {
  username: string;
  password: string;
  role: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  cashier: 'cashier',
} as const;

export interface UserUpdate {
  password?: string;
  role?: UserUpdateRole;
}

export type CourtStatus = typeof CourtStatus[keyof typeof CourtStatus];


export const CourtStatus = {
  idle: 'idle',
  active: 'active',
} as const;

export interface Court {
  id: number;
  name: string;
  hourlyRate: number;
  status: CourtStatus;
  /** @nullable */
  activeSessionId?: number | null;
  /** @nullable */
  activeSessionStartedAt?: string | null;
  /** @nullable */
  activeSessionElapsedMinutes?: number | null;
  /** @nullable */
  activeSessionCurrentCost?: number | null;
}

export interface CourtInput {
  name: string;
  hourlyRate: number;
}

export interface CourtUpdate {
  name?: string;
  hourlyRate?: number;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export interface Session {
  id: number;
  courtId: number;
  courtName: string;
  startedAt: string;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  totalCost?: number | null;
  status: SessionStatus;
}

export interface SessionInput {
  courtId: number;
  /** @nullable */
  presetMinutes?: number | null;
}

export interface Expense {
  id: number;
  type: string;
  amount: number;
  date: string;
}

export interface ExpenseInput {
  type: string;
  amount: number;
  date?: string;
}

export interface ReportSummary {
  totalIncome: number;
  totalExpenses: number;
  netProfit: number;
  sessions: Session[];
  expenses: Expense[];
}

export interface DashboardStats {
  totalActiveCourts: number;
  totalIdleCourts: number;
  todayIncome: number;
  activeSessions: number;
}

export interface Settings {
  id: number;
  systemName: string;
  themeColor: string;
  /** @nullable */
  telegramApiKey?: string | null;
  /** @nullable */
  discordWebhookUrl?: string | null;
}

export interface SettingsUpdate {
  systemName?: string;
  themeColor?: string;
  telegramApiKey?: string;
  discordWebhookUrl?: string;
}

export type GetSessionsParams = {
date?: string;
courtId?: number;
};

export type GetExpensesParams = {
date?: string;
};

export type GetReportSummaryParams = {
date?: string;
startDate?: string;
endDate?: string;
};