# lib/api-zod/src/generated/types/court.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tennis Ranya API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CourtStatus } from './courtStatus';

export interface Court {
  id: number;
  name: string;
  hourlyRate: number;
  status: CourtStatus;
  /** @nullable */
  activeSessionId?: number | null;
  /** @nullable */
  activeSessionStartedAt?: string | null;
  /** @nullable */
  activeSessionElapsedMinutes?: number | null;
  /** @nullable */
  activeSessionCurrentCost?: number | null;
}