# lib/api-zod/src/generated/types/courtInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tennis Ranya API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourtInput {
  name: string;
  hourlyRate: number;
}