# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tennis Ranya API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  courtId: number;
  /** @nullable */
  presetMinutes?: number | null;
}