# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tennis Ranya API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  systemName: string;
  themeColor: string;
  /** @nullable */
  telegramApiKey?: string | null;
  /** @nullable */
  discordWebhookUrl?: string | null;
}